Define custom batch type
