+++ example/src/main/scala/example/Contributors.scala
@@ -1,55 +1,11 @@
 package example
 
-import example.Contributor._
-import example.ApiError._
-import io.circe.parser.decode
-import io.circe.syntax._
-import sttp.client.asynchttpclient.zio.SttpClient
-import sttp.client.circe.asJson
-import sttp.client.{ UriContext, basicRequest }
-import sttp.model.Uri
+import io.circe.generic.semiauto._
+import zio.Chunk
+import io.circe.Encoder
 
-import zio._
-import zio.duration._
-import zio.redis._
+final case class Contributors(contributors: Chunk[Contributor]) extends AnyVal
 
 object Contributors {
-
-  trait Service {
-    def fetchContributors(organization: String, repository: String): IO[ApiError, Chunk[Contributor]]
-  }
-
-  lazy val live: ZLayer[RedisExecutor with SttpClient, Nothing, Contributors] =
-    ZLayer.fromFunction { env =>
-      new Service {
-        def fetchContributors(organization: String, repository: String): IO[ApiError, Chunk[Contributor]] = {
-          val key = s"$organization:$repository"
-          sMembers(key).flatMap { response =>
-            if (response.isEmpty)
-              for {
-                contributors <- retrieveContributors(organization, repository)
-                _            <- sAdd(key, contributors.asJson.toString, contributors.map(_.asJson.toString): _*)
-                _            <- pExpire(key, 1.minute)
-              } yield contributors
-            else
-              response
-                .mapM(contributor => ZIO.fromEither(decode[Contributor](contributor)))
-                .orElseFail(GithubUnavailable)
-          }.orElseFail(GithubUnavailable).provide(env)
-        }
-      }
-    }
-
-  private[this] def retrieveContributors(
-    organization: String,
-    repository: String
-  ): ZIO[SttpClient, ApiError, Chunk[Contributor]] =
-    SttpClient
-      .send(basicRequest.get(urlOf(organization, repository)).response(asJson[Chunk[Contributor]]))
-      .flatMap(_.body.fold(_ => ZIO.fail(UnknownProject), ZIO.succeed(_)))
-      .orElseFail(GithubUnavailable)
-
-  private[this] def urlOf(organization: String, repository: String): Uri =
-    uri"https://api.github.com/repos/$organization/$repository/contributors"
-
+  implicit val encoder: Encoder[Contributors] = deriveEncoder[Contributors]
 }
