+++ example/src/main/scala/example/package.scala
@@ -1,5 +1,5 @@
 import zio.Has
 
 package object example {
-  type Contributors = Has[Contributors.Service]
+  type ContributorsCache = Has[ContributorsCache.Service]
 }
